+++ integrations/tailwindcss-cli/tests/cli.test.js
@@ -2,6 +2,7 @@ let path = require('path')
 let $ = require('../../execute')
 let { css, html, javascript } = require('../../syntax')
 let resolveToolRoot = require('../../resolve-tool-root')
+let { env } = require('../../../lib/lib/sharedState')
 
 let version = require('../../../package.json').version
 
@@ -215,22 +216,42 @@ describe('Build command', () => {
 
     await $(`${EXECUTABLE} --output ./dist/main.css --postcss`)
 
-    expect(await readOutputFile('main.css')).toIncludeCss(
-      css`
-        .font-bold-after {
-          font-weight: 700;
-        }
+    if (!env.OXIDE) {
+      expect(await readOutputFile('main.css')).toIncludeCss(
+        css`
+          .font-bold-after {
+            font-weight: 700;
+          }
 
-        .btn-after {
-          --tw-bg-opacity: 1;
-          background-color: rgb(239 68 68 / var(--tw-bg-opacity));
-          padding-left: 0.5rem;
-          padding-right: 0.5rem;
-          padding-top: 0.25rem;
-          padding-bottom: 0.25rem;
-        }
-      `
-    )
+          .btn-after {
+            --tw-bg-opacity: 1;
+            background-color: rgb(239 68 68 / var(--tw-bg-opacity));
+            padding-left: 0.5rem;
+            padding-right: 0.5rem;
+            padding-top: 0.25rem;
+            padding-bottom: 0.25rem;
+          }
+        `
+      )
+    }
+
+    if (env.OXIDE) {
+      expect(await readOutputFile('main.css')).toIncludeCss(
+        css`
+          .font-bold-after {
+            font-weight: 700;
+          }
+
+          .btn-after {
+            background-color: #ef4444;
+            padding-left: 0.5rem;
+            padding-right: 0.5rem;
+            padding-top: 0.25rem;
+            padding-bottom: 0.25rem;
+          }
+        `
+      )
+    }
   })
 
   test('--postcss (custom.postcss.config.js)', async () => {
@@ -266,22 +287,42 @@ describe('Build command', () => {
 
     await $(`${EXECUTABLE} --output ./dist/main.css --postcss ./custom.postcss.config.js`)
 
-    expect(await readOutputFile('main.css')).toIncludeCss(
-      css`
-        .font-bold-after {
-          font-weight: 700;
-        }
+    if (!env.OXIDE) {
+      expect(await readOutputFile('main.css')).toIncludeCss(
+        css`
+          .font-bold-after {
+            font-weight: 700;
+          }
 
-        .btn-after {
-          --tw-bg-opacity: 1;
-          background-color: rgb(239 68 68 / var(--tw-bg-opacity));
-          padding-left: 0.5rem;
-          padding-right: 0.5rem;
-          padding-top: 0.25rem;
-          padding-bottom: 0.25rem;
-        }
-      `
-    )
+          .btn-after {
+            --tw-bg-opacity: 1;
+            background-color: rgb(239 68 68 / var(--tw-bg-opacity));
+            padding-left: 0.5rem;
+            padding-right: 0.5rem;
+            padding-top: 0.25rem;
+            padding-bottom: 0.25rem;
+          }
+        `
+      )
+    }
+
+    if (env.OXIDE) {
+      expect(await readOutputFile('main.css')).toIncludeCss(
+        css`
+          .font-bold-after {
+            font-weight: 700;
+          }
+
+          .btn-after {
+            background-color: #ef4444;
+            padding-left: 0.5rem;
+            padding-right: 0.5rem;
+            padding-top: 0.25rem;
+            padding-bottom: 0.25rem;
+          }
+        `
+      )
+    }
   })
 
   test('--postcss supports process options', async () => {