+++ tests/prefers-contrast.test.js
@@ -1,6 +1,6 @@
 import { crosscheck, run, html, css, defaults } from './util/run'
 
-crosscheck(() => {
+crosscheck(({ stable, oxide }) => {
   it('should be possible to use contrast-more and contrast-less variants', () => {
     let config = {
       content: [
@@ -18,7 +18,7 @@ crosscheck(() => {
     `
 
     return run(input, config).then((result) => {
-      expect(result.css).toMatchFormattedCss(css`
+      stable.expect(result.css).toMatchFormattedCss(css`
         ${defaults}
         .bg-white {
           --tw-bg-opacity: 1;
@@ -34,6 +34,22 @@ crosscheck(() => {
           .contrast-less\:bg-black {
             --tw-bg-opacity: 1;
             background-color: rgb(0 0 0 / var(--tw-bg-opacity));
+          }
+        }
+      `)
+      oxide.expect(result.css).toMatchFormattedCss(css`
+        ${defaults}
+        .bg-white {
+          background-color: #fff;
+        }
+        @media (prefers-contrast: more) {
+          .contrast-more\:bg-pink-500 {
+            background-color: #ec4899;
+          }
+        }
+        @media (prefers-contrast: less) {
+          .contrast-less\:bg-black {
+            background-color: #000;
           }
         }
       `)
@@ -53,7 +69,7 @@ crosscheck(() => {
     `
 
     return run(input, config).then((result) => {
-      expect(result.css).toMatchFormattedCss(css`
+      stable.expect(result.css).toMatchFormattedCss(css`
         ${defaults}
         @media (prefers-contrast: more) {
           .contrast-more\:bg-black {
@@ -65,6 +81,19 @@ crosscheck(() => {
           .dark\:bg-white {
             --tw-bg-opacity: 1;
             background-color: rgb(255 255 255 / var(--tw-bg-opacity));
+          }
+        }
+      `)
+      oxide.expect(result.css).toMatchFormattedCss(css`
+        ${defaults}
+        @media (prefers-contrast: more) {
+          .contrast-more\:bg-black {
+            background-color: #000;
+          }
+        }
+        @media (prefers-color-scheme: dark) {
+          .dark\:bg-white {
+            background-color: #fff;
           }
         }
       `)